refactor: 최단 거리 반환 타입 변경(double -> int)

+++ src/main/java/wooteco/subway/path/domain/SubwayMap.java
@@ -41,8 +41,8 @@ public class SubwayMap {
         stations.forEach(subwayMap::addVertex);
     }
 
-    public double getShortestPath(Station source, Station target) {
-        return map.getPathWeight(source, target);
+    public int getShortestPath(Station source, Station target) {
+        return (int) map.getPathWeight(source, target);
     }
 
     public List<Station> getStationsOnPath(Station source, Station target) {
